Removed Bill making problem
There was a problem with the bill fuction. When user entered quantity of product is greater than product quantity the bill consisted of that product price even if the user doesn't want to buy them. This problem was fixed and also when the product quantity is zero then we print that the product is out of stock